requirements.txt
allure-pytest==2.8.0
allure-python-commons==2.8.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Faker==2.0.1
idna==2.8
importlib-metadata==0.19
jsonschema==3.0.2
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.1
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
text-unidecode==1.2
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
